questdb-sink/requirements.txt
quixstreams[influxdb3]==3.16.1
python-dotenv
quixstreams
psycopg2-binary
